feat: move str_to_cstr_bytes function to utils module and update imports
